start moving methods to the best location
I am grouping them by the functionalities and place them in the best file. config should be used across all of them.
