+++ automation/gui.py
@@ -1,0 +1,3 @@
+"""
+A collection of pyautogui method
+"""
