--- automation/pazusoba.py
@@ -1,5 +1,0 @@
-"""
-pazusoba automation
-
-Working in progress...
-"""
